improve apikey delete

+++ src/models/ApiKeys.php
@@ -63,12 +63,13 @@ class ApiKeys implements RestInterface
 
     /**
      * Create a known key so we can test against it in dev mode
+     * It can also be used to create an initial sysadmin key
      * This function should only be called from the db:populate command
      */
     public function createKnown(string $apiKey): int
     {
         $hash = password_hash($apiKey, PASSWORD_BCRYPT);
-        return $this->insert('known key used for tests', 1, $hash);
+        return $this->insert('known key used from db:populate command', 1, $hash);
     }
 
     /**
@@ -126,9 +127,10 @@ class ApiKeys implements RestInterface
 
     public function destroy(): bool
     {
-        $sql = 'DELETE FROM api_keys WHERE id = :id';
+        $sql = 'DELETE FROM api_keys WHERE id = :id AND userid = :userid';
         $req = $this->Db->prepare($sql);
         $req->bindValue(':id', $this->id, PDO::PARAM_INT);
+        $req->bindValue(':userid', $this->Users->requester->userid ?? 0, PDO::PARAM_INT);
 
         if ($res = $this->Db->execute($req)) {
             AuditLogs::create(new ApiKeyDeleted($this->Users->requester->userid ?? 0, $this->Users->userid ?? 0));
